requirements.txt
# Core dependencies
tabulate==0.9.0
requests==2.31.0

# ML and data processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
mlflow==2.6.0

# API and web framework
fastapi==0.100.0
uvicorn==0.23.0
pydantic==2.0.2

# Testing and quality
pytest==7.4.0
pytest-cov==4.1.0
flake8==6.0.0

# Data versioning
dvc==3.9.0
dvc[s3]==3.9.0

# Dev tools
black==23.3.0
isort==5.12.0

# YAML processing
pyyaml==6.0 